leetcode/src/main/java/org/example/graph/CloneGraph.java
package org.example.graph;

import java.util.HashMap;
import java.util.Map;

public class CloneGraph {
    private Map<Node, Node> visited = new HashMap<>();
    public Node cloneGraph(Node node) {
        if (node == null) {
            return null;
        }
        if (visited.containsKey(node)) {
            return visited.get(node);
        }
        Node clonedNode = new Node(node.val);
        visited.put(node, clonedNode);
        for (Node neighbor : node.neighbors) {
            clonedNode.neighbors.add(cloneGraph(neighbor));
        }
        return clonedNode;
    }
}
